refactor CTE_map from SelectStatement to QueryNode
