+++ src/include/duckdb/parser/query_node/select_node.hpp
@@ -13,6 +13,7 @@
 #include "duckdb/parser/sql_statement.hpp"
 #include "duckdb/parser/tableref.hpp"
 #include "duckdb/parser/parsed_data/sample_options.hpp"
+#include "duckdb/parser/statement/select_statement.hpp"
 
 namespace duckdb {
