+++ src/include/duckdb/planner/expression_binder.hpp
@@ -99,7 +99,6 @@ protected:
 	                             unique_ptr<ParsedExpression> *expr_ptr);
 
 	virtual void ReplaceMacroParametersRecursive(unique_ptr<ParsedExpression> &expr);
-	void ReplaceMacroParametersRecursive(ParsedExpression &expr, SelectStatement &statement);
 	virtual void ReplaceMacroParametersRecursive(ParsedExpression &expr, QueryNode &node);
 	virtual void ReplaceMacroParametersRecursive(ParsedExpression &expr, TableRef &ref);
 	virtual void CheckForSideEffects(FunctionExpression &function, idx_t depth, string &error);
