+++ src/planner/binder/expression/bind_macro_expression.cpp
@@ -24,7 +24,7 @@ void ExpressionBinder::ReplaceMacroParametersRecursive(unique_ptr<ParsedExpressi
 	case ExpressionClass::SUBQUERY: {
 		// replacing parameters within a subquery is slightly different
 		auto &sq = ((SubqueryExpression &)*expr).subquery;
-		ReplaceMacroParametersRecursive(*expr, *sq);
+		ReplaceMacroParametersRecursive(*expr, *sq->node);
 		break;
 	}
 	default: // fall through
@@ -33,13 +33,6 @@ void ExpressionBinder::ReplaceMacroParametersRecursive(unique_ptr<ParsedExpressi
 	// unfold child expressions
 	ParsedExpressionIterator::EnumerateChildren(
 	    *expr, [&](unique_ptr<ParsedExpression> &child) { ReplaceMacroParametersRecursive(child); });
-}
-
-void ExpressionBinder::ReplaceMacroParametersRecursive(ParsedExpression &expr, SelectStatement &statement) {
-	ReplaceMacroParametersRecursive(expr, *statement.node);
-	for (auto &kv : statement.cte_map) {
-		ReplaceMacroParametersRecursive(expr, *kv.second->query);
-	}
 }
 
 void ExpressionBinder::ReplaceMacroParametersRecursive(ParsedExpression &expr, TableRef &ref) {
@@ -68,7 +61,7 @@ void ExpressionBinder::ReplaceMacroParametersRecursive(ParsedExpression &expr, T
 	}
 	case TableReferenceType::SUBQUERY: {
 		auto &sq_ref = (SubqueryRef &)ref;
-		ReplaceMacroParametersRecursive(expr, *sq_ref.subquery);
+		ReplaceMacroParametersRecursive(expr, *sq_ref.subquery->node);
 		break;
 	}
 	case TableReferenceType::TABLE_FUNCTION: {
@@ -119,6 +112,9 @@ void ExpressionBinder::ReplaceMacroParametersRecursive(ParsedExpression &expr, Q
 	}
 	default:
 		throw NotImplementedException("QueryNode type not implemented for macro's!");
+	}
+	for (auto &kv : node.cte_map) {
+		ReplaceMacroParametersRecursive(expr, *kv.second->query->node);
 	}
 }
