+++ src/planner/binder/statement/bind_select.cpp
@@ -6,11 +6,6 @@ namespace duckdb {
 
 BoundStatement Binder::Bind(SelectStatement &stmt) {
 	this->allow_stream_result = true;
-	// first we visit the set of CTEs and add them to the bind context
-	for (auto &cte_it : stmt.cte_map) {
-		AddCTE(cte_it.first, cte_it.second.get());
-	}
-	// now visit the root node of the select statement
 	return Bind(*stmt.node);
 }
