+++ src/planner/binder/tableref/bind_subqueryref.cpp
@@ -10,9 +10,6 @@ unique_ptr<BoundTableRef> Binder::Bind(SubqueryRef &ref, CommonTableExpressionIn
 		binder->bound_ctes.insert(cte);
 	}
 	binder->alias = ref.alias;
-	for (auto &cte_it : ref.subquery->cte_map) {
-		binder->AddCTE(cte_it.first, cte_it.second.get());
-	}
 	auto subquery = binder->BindNode(*ref.subquery->node);
 	idx_t bind_index = subquery->GetRootIndex();
 	auto result = make_unique<BoundSubqueryRef>(move(binder), move(subquery));
